GC 118: I:ZI Miner handles error when DataDescription is not available

+++ tools/IZI Miner/web/js/lib/Config.js
@@ -64,7 +64,7 @@ var Config = new Class({
 	},
 
     getIdDm: function() {
-        return this.params.id_dm.toInt();
+        return this.params.id_dm;
     },
 	
 	getBKAskURL: function () {
